First attempt at a simple mcp
Initial attempt will just report based on exit code of Orion. To
determine a regression, it won't use a single configuration file.

Eventually it would be cool to have other products, so from this mcp
servre we could ask in general terms -- "Is the Red Hat Portfolio that
Red Hat Performance covers regressing in Q2?"

Signed-off-by: Joe Talerico aka rook <joe.talerico@gmail.com>